backend/requirements.txt
fastapi[standard]
python-dotenv
odmantic